refactorize for reduce size
